+++ src/search/move_orderer.rs
@@ -1010,8 +1010,8 @@ mod tests {
             results.tags_mut().remove(Tag::ACN);
             println!(
                 "{:>12} {:>12} {}",
-                Formatting::u128(nodes),
-                Formatting::u128(nodes_cumul),
+                Formatting::u64(nodes),
+                Formatting::u64(nodes_cumul),
                 results
             );
         }
